django_langchain_api_example/requirements.txt
django
django-cors-headers
djangorestframework
ipython
langchain
openai
pip-chill
python-dotenv
